+++ segretini_matplottini/plot/__init__.py
@@ -1,5 +1,7 @@
 from segretini_matplottini.plot.barplot import (  # noqa: F401
+    barplot,
     barplot_for_multiple_categories,
+    barplots,
 )
 from segretini_matplottini.plot.correlation_scatterplot import (
     correlation_scatterplot,  # noqa: F401